requirements.txt
accessify==0.3.1
aker==0.5.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
awesome_module==0.1.0
bcrypt==4.1.3
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
configparser==7.0.0
coverage==7.5.1
cryptography==42.0.5
DateTime==5.5
Faker==25.2.0
GitHub-Action-Python-Example==1.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
importlib2==3.5.0.2
iniconfig==2.0.0
Jinja2==3.1.4
kaitaistruct==0.10
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
paramiko==3.4.0
pluggy==1.5.0
pyasn1==0.5.1
pycparser==2.21
pydivert==2.1.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
pyte==0.7.0
pytest==8.2.0
pytest-base-url==2.1.0
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-timeout==2.3.1
pytest-variables==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
selenium==4.18.1
selenium-webdriver-extender==0.0.0
selenium-wire==5.1.0
selenium_web_browser==0.0.2
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
start-selenium-webdriver==0.3.1
tenacity==8.3.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
urwid==2.6.11
wait==0.0.3
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
